backend/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "agent-studio-starter Python backend"
#readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "copilotkit>=0.1.78",
    "deepagents==0.3.12",
    "fastapi>=0.115.14",
    "langchain>=1.2.9",
    "langchain-litellm==0.4.0",
    "langchain-openai>=1.1.8",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "tavily-python>=0.7.21",
    "uvicorn[standard]>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist.force-include]
"../src/agent" = "src/agent"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[dependency-groups]
dev = [
    "ruff==0.15.0",
    "pytest==9.0.2",
    "mypy==1.19.1",
    "deepeval==3.8.4",
    "ipykernel==7.2.0",
    "ipywidgets==8.1.8",
]

[tool.mypy]
python_version = '3.13'
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
follow_imports = 'skip'
ignore_missing_imports = true
show_column_numbers = true
show_error_codes = true
warn_unused_ignores = true
warn_return_any = true

[tool.ruff]
exclude = [
    "scripts",
    "notebooks"
]
line-length = 100
target-version = "py313"
fix = true
lint.select = [
    # pycodestyle E
    "E4", "E7", "E9", 
    # Pyflakes
    "F", 
    # pycodestyle W
    "W",
    # isort 
    "I",
    # mccabe
    "C90",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # pylint
    "PL",
    # flake8-bugbear
    "B006",
]

